+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 **v4.0**
   - Extract routing mechanism into a `Router` class
   - Clean `ServiceBroker` interface
+  - `gevent` is now optional and can be installed by `pip install openbrokerapi[gevent]`
 
 **Incompatibility**
   - Changed parameter order in UnbindDetails, DeprovisionDetails